Issue #3 Build for EL5
Massive changes to SPEC file to work with EL5, jk!

I added an if statement towards the top, and made it look similar to
other python26 modules.

I also cleaned up some lint errors, the only lint error right now is
for the URL on Source0. We probably need to think about semantic
versioning, and maybe upload the source file with the proper name?

~tommy